+++ packages/renderer/src/hooks/mask.ts
@@ -1,8 +1,9 @@
 import { ref, watchEffect } from 'vue';
 import { useRoute } from 'vue-router';
-import { type MenuDataItem } from '@vtj/ui';
-import { type PageFile } from '@vtj/core';
+import type { MenuDataItem } from '@vtj/ui';
+import type { PageFile } from '@vtj/core';
 import { useProvider } from '../provider';
+import { useAccess } from '../plugins';
 
 function createMenus(pages: PageFile[] = []): MenuDataItem[] {
   return pages.map((page) => {
@@ -19,10 +20,36 @@ function createMenus(pages: PageFile[] = []): MenuDataItem[] {
   });
 }
 
+function menusFilter(
+  menus: MenuDataItem[],
+  permissions?: Record<string, boolean>
+): MenuDataItem[] {
+  if (!permissions) return menus;
+  let result: MenuDataItem[] = [];
+
+  for (const menu of menus) {
+    if (menu.children && menu.children.length) {
+      const childArray = menusFilter(menu.children, permissions);
+      if (childArray.length) {
+        menu.children = childArray;
+        result.push(menu);
+      }
+    } else {
+      if (permissions[menu.id]) {
+        result.push(menu);
+      }
+    }
+  }
+
+  return result;
+}
+
 export function useMask() {
   const provider = useProvider();
   const route = useRoute();
+  const access = useAccess();
   const disabled = ref(false);
+  const accessData = access?.getData();
   const project = provider.project;
   watchEffect(() => {
     const { name, params, meta } = route;
@@ -44,6 +71,6 @@ export function useMask() {
     logo: config?.logo,
     themeSwitchable: config?.themeSwitchable,
     title: config?.title || project?.description || project?.name || 'VTJ App',
-    menus
+    menus: menusFilter(menus, accessData?.permissions)
   };
 }